+++ example/midway-react-ssr/typings/config/plugin.d.ts
@@ -18,18 +18,18 @@ import 'midway-schedule';
 import { EggPluginItem } from 'egg';
 declare module 'egg' {
   interface EggPlugin {
-    onerror?: EggPluginItem;
-    session?: EggPluginItem;
-    i18n?: EggPluginItem;
-    watcher?: EggPluginItem;
-    multipart?: EggPluginItem;
-    security?: EggPluginItem;
-    development?: EggPluginItem;
-    logrotator?: EggPluginItem;
-    schedule?: EggPluginItem;
-    static?: EggPluginItem;
-    jsonp?: EggPluginItem;
-    view?: EggPluginItem;
-    schedulePlus?: EggPluginItem;
+    'onerror'?: EggPluginItem;
+    'session'?: EggPluginItem;
+    'i18n'?: EggPluginItem;
+    'watcher'?: EggPluginItem;
+    'multipart'?: EggPluginItem;
+    'security'?: EggPluginItem;
+    'development'?: EggPluginItem;
+    'logrotator'?: EggPluginItem;
+    'schedule'?: EggPluginItem;
+    'static'?: EggPluginItem;
+    'jsonp'?: EggPluginItem;
+    'view'?: EggPluginItem;
+    'schedulePlus'?: EggPluginItem;
   }
 }