+++ example/midway-vue3-ssr/vite.server.config.js
@@ -14,7 +14,6 @@ module.exports = {
     middlewareMode: 'ssr'
   },
   define: {
-    __isVite__: true,
     __isBrowser__: false
   },
   resolve: {