+++ example/midway-vue3-ssr/web/components/layout/App.vue
@@ -5,6 +5,8 @@
 <script lang="ts">
 // 在这里可以进行一些全局组件的注册逻辑
 export default {
-
+  created () {
+    console.log('xxx', __isBrowser__)
+  }
 }
 </script>